movepid: send a plain int pid
Detect whether the caller is in the same pidns as us.  If so, then
we accept the victim pid as a plain integer.  If not, then the
caller must send the pid as an SCM_CREDENTIAL so we can convert
it.

We jump through these hoops because non-root users cannot send
SCM_CREDENTIAL pids for tasks other than themselves.  A root-owned
proxy in their own pidns can convert the  int pid to an SCM_CREDENTIAL
for us.  But for performance's sake we'd rather not do that for the
initial pidns.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>