fixes to chown
move the chown helper to fs.c

when requestor r requests cgroup /a/b/c to be created, then we chown
all files under /a/b/c to r's uid/gid.  This is because r already had
the rights to write under /a/b, therefore writing to files under
/a/b/c/ is safe.  If r requests chowning /a/b/c to some uid (its own
or any other), then only the directory and proc+tasks files are
chowned.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>