add scm versions of all methods
In these, the proxy passes the requestor's credentials as an
scm_cred.  The proxy could not do so if it did not have privilege
over the passed creds.

In this way, the proxy doesn't need to do any checking of the
requestor permissions at all, instead deferring wholly to the
manager.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>