move cgmanager socket into a private directory
The issue is:  when a container is started, it gets a bind mount
of the socket itself.  If the manager exits and restarts, then
the container must be stopped and re-started to pick up the new
bind mount.

So keep a /sys/fs/cgroup/cgmanager/ directory and bind mount
that into the container instead.  The cgproxy moves the directory
mount to /sys/fs/cgroup/cgmanager.lower and creates its own
/sys/fs/cgroup/cgmanager.  Now if the cgmanager restarts,
the cgproxy in the container can detect that and simply reload
the new socket from the /sys/fs/cgroup/cgmanager directory.
Note cgproxy does not yet detect this condition on its own, but
that should be easy to do.  At least now the container can simply
stop cgproxy;  start cgproxy and pick up the new socket.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>