create: allow caller to determine if cgroup already existed
Create() now takes an output parameter so the caller can determine if
the group already existed. The scm variants now return '1' to denote
successful creation or '2' to denote that the group already existed.

create_main() will now not call chown if the group already existed.

[ Serge: modified to not be on top of Exclusive calls, and modified
to hopefully DTRT with paths longer than 1 directory ]

Signed-off-by: James Hunt <james.hunt@ubuntu.com>
Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>