cgmanager_chown: pass a valid pid
The cgmanager ignores the victim pid passed to it, but the proxy has to
pass it along to the manager in an scm cred.  If proxy tries to pass an
scm_cred with pid 0, the kernel will refuse as the pid is not valid.  So
just send our own pid.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>