cgm-concurrent: add connect-only mode
running cgm-concurrent -i 10 -j 2 -c will *only* actually
connect and disconnect from the daemon.  This sadly still
can segfault (or, on a faster system, -j 30 -i 50), showing
that there is a threading bug in dbus (maybe nih-dbus), or
we are misunderstanding how to properly use it (and so doing
something wrong).

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>