implement release-on-empty
The new kernel mechanism for automatic removal of empty cgroups
is not yet available, therefore this uses the original
release-agent feature.  The release-agent is only told the
cgroup name being deleted, not the controller.

Cgmanager installs a program cgm-release-agent into /sbin.  It
then symlinks that to cgm-release-agent.$controller under
/run/cgmanager/agents and registers each symlink with the
related controller.  The RemoveOnEmpty dbus call causes 1 to
be written to the notify-on-release file for the specified
cgroup, so that when it becomes empty the
cgm-release-agent.$controller will be called.

As a part of this the mounting of subsystems is broken up into
collecting the subsystems, then setting release agents, and
finally mounting the subsystems into our private namespace.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>