add timeout to connections
To avoid having lazy callers fill up our fdtable.  Every new method that
is called causes the timeout to be reset.

In order to avoid having to walk all of the nih_timers to look for the
one to reset, we keep our own separate hash of connection timers keyed
on the connection itself.  This is the bulk of the new code.

Also increase our open file limit to 10k.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>