Implement getpidcgroupabs
This returns the cgroup path of another task relative to the
cgproxy if used, or else the full path as seen by the cgmanager.

This path can then be fed to movepidabs.  This can be used by
lxc-attach to re-attach to a container's cgroup even if the
user's new login shell is in a different cgroup session than
that under which the container was started.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>