do_gettasks: detect 0 length array
the nih-dbus-tool-generated library raises an error if we return an empty
array, even though we returned 0 (success) in the dbus service.  So
detect whether pids_len was changed to 0, and if so just suck up the
error, print nothing (empty set) and return success.

Also add newlines to a bunch of error messages.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>