cgm: ignore leading lt- in basename for usage
In package builds, ./cgm is a libtool wrapper around ./.libs/cgm.  In
jenkins, it is not.  When ./cgm is a wrapper, then argv[0] is "lt-cgm".
That confuses help2man, so I tried to have help2man run ./.libs/cgm to
get the usage output, but that breaks in jenkins.

So just have cgm check whether it is called as "lt-foo", and call itself
foo in usage output if so.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>