get_tasks: don't hang proxy if tasks are exiting
In both GetTasks and GetTasksRecursive in the proxy, switch to a 1-second
select plus recvmsg(MSG_DONTWAIT) when waiting for pids from the cgmanager.
if the cgmanager sees a credential send failed, then it will resend the
first pid which succeeded, so that the proxy just gets a dup.  If no previous
pid has succeeded, or if that pid has now also exited, then we'll simply exit,
and the proxy will now gracefully fail.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>