fix: force libnih to use /run for pidfile
After unshare() and pivot_root() during daemonization, we do not take
/var in with us (only /, /proc, and /run)

Since no /var exists after pivot, this means /var/run will not exist if
it's on a separate mount point from /.  On contemporary systems for a
long while, /var/run is either a symlink or bind mount to /run anyways,
so we just tell libnih explicitly to use the pidfile in there rather
than using its default /var/run path (which won't work if /var is
separate anyways).

Patch originally submitted by rworkman@slackware.com, but they decided
to change libnih default to /run universally, so this patch wouldn't be
needed in that case (but wouldn't be a problem either).

Closes #22