fix: sysvinit invoke via path lookup, not /sbin
The sysvinit scripts assume the install is in /sbin, and hardcode that
as the invocation path, even though the 'configure' script itself
defaults to /usr/local if no --prefix override is given (as standard for
autoconf).

This means an out-of-box compile will not even work with the provided
init scripts without modifying them by hand.

The locations used are:

  - /usr/local/sbin/cgmanager: installation default
  - /usr/bin/cgmanager: gentoo scripts assumption
  - /sbin/cgmanager: debian packages give this prefix (and this is what
                     the script hardcodes)

Instead of harcoding one of the paths, have the init script look up the
invocation path at runtime to invoke cgmanager (although we still use a
hardcoded PATH, but include all three locations).