fix: sysvinit: umount cgmfs on stop, avoid pileup
cgmanager mounts cgmfs on /run/cgmanager/fs, but does not clean it up on
exit, so they pile up additional ones every stop/start cycle unless we
try to unmount it on stop.

this only attempts to fix sysvinit script.  upstart and systemd should be
tested as well, maybe it does not matter for them.

workaround for #26