Make the PAM_CGM module more configurable
This commit cleans up the PAM_CGM module and makes it much more
configurable.

First, the module is rewritten to keep its data using a PAM module internal
data mechanism instead of using a set of static variables since PAM API
allows doing multiple transactions in parallel, including utilizing the
same module, as long as each one uses a separate transaction handle.

Additionally, a PAM man page for pam_set_data() specifically says that
"[i]n general such files [PAM modules] should not contain static
variables".

Second, the following configuration options are added:
* A pattern setting, which allows naming cgroups created by this module
in a different way from "user/foo/0",

* An ability to disable appending of an index suffix to a pattern-derived
cgroup name,

* A limit setting which allows controlling how large these suffix indices
(if they aren't disabled by the previous option) are allowed to grow,

* An ability to have particular values set in the created final cgroups,

* A setting controlling how deep cgroups prune process goes at the session
close time.

All these options (and their behavior) are described at the top of the
pam_cgm.c file, where there was already a description of the module
operation.

Unfortunately, we cannot use libnih option parsing since it does not work
in a shared library context so in order not to reinvent the wheel the
popular popt library was used instead.

The default values of all these options were set in a way so not to result
in any change of the module behavior, unless specifically configured in a
different way by an user, for compatibility with existing deployments.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>