Add an ability to allow autoremove on pre-mounted v1 controllers
Some init systems (for example OpenRC) mount v1 controllers on boot to be
able to put their services under them but do not set any release agent on
these mounts.
If such controller is then used with cgmanager the autoremove (or
remove-on-empty) functionality will not work for this controller.

This means that when, for example, per-user-session cgroups are created
under such controller, too, empty cgroups from past sessions will litter
the system.

It is safe to allow autoremove and use our release agent for such
pre-mounted controllers, as long as we enable it (via "notify_on_release")
only on particular (sub-)cgroups that we had received an autoremove request
for.

Add two new options to cgmanager to possibly allow just that (they default
to "disabled" setting for backwards compatibility).

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>